feat: add deal status command
